+++ geo/src/test/java/com/tomtom/speedtools/geometry/GeoPointTest.java
@@ -80,15 +80,15 @@ public class GeoPointTest {
     @Test
     public void testWithElevationMeters() {
         LOG.info("testWithElevationMeters");
-        Assert.assertEquals(0, Double.compare(y2, x.withElevationMeters(y2).getElevationMeters()));
-        Assert.assertEquals(0, Double.compare(x2, y.withElevationMeters(x2).getElevationMeters()));
+        Assert.assertEquals(0, Double.compare(y2, x.withElevationMeters(y2).getElevationMetersOrNaN()));
+        Assert.assertEquals(0, Double.compare(x2, y.withElevationMeters(x2).getElevationMetersOrNaN()));
     }
 
     @Test
     public void testWithElevationMetersOrNull() {
         LOG.info("testWithElevationMeters");
-        Assert.assertEquals(0, Double.compare(y2, x.withElevationMeters(y2).getElevationMetersOrNull()));
-        Assert.assertEquals(0, Double.compare(x2, y.withElevationMeters(x2).getElevationMetersOrNull()));
+        Assert.assertEquals(0, Double.compare(y2, x.withElevationMeters(y2).getElevationMeters()));
+        Assert.assertEquals(0, Double.compare(x2, y.withElevationMeters(x2).getElevationMeters()));
     }
 
     @Test