+++ src/site/markdown/ReleaseNotes.md
@@ -4,7 +4,7 @@ Release Notes
 ### 3.2.5
 
 * Included `elevationMeters` in `Geo`-classes, so `GeoPoint`s can be 3D. If the elevation is omitted, the
-point is considered 2D.
+point is considered 2D. 
 
 * Includes moving to Scala 2.12.4 and Akka libraries compiled for Scala 2.12.