+++ src/site/markdown/ReleaseNotes.md
@@ -1,6 +1,12 @@
 Release Notes
 ----
 
+### 3.2.5
+
+* Included `elevationMeters` in `GeoPoint`, so points can be 3D.
+
+* Includes moving to Scala 2.12.4 and Akka libraries compiled for Scala 2.12.
+
 ### 3.2.4
 
 * Updated dependencies to latest from Maven Central.